+++ src/runtime/gc-typedefs.h
@@ -3,6 +3,6 @@
 /* These are for grovel-headers because depending on how pedantic your system linker is,
  * it might be an error to try to include gc.h and then not link in any file that
  * provides the inline functions, even if they're never used. */
-typedef intptr_t page_index_t;
+typedef int page_index_t;
 typedef signed char generation_index_t;
 #endif
